+++ cvat/apps/auto_annotation/views.py
@@ -3,32 +3,35 @@
 #
 # SPDX-License-Identifier: MIT
 
-from django.http import HttpResponse, JsonResponse, HttpResponseBadRequest
-from rules.contrib.views import permission_required, objectgetter
-from cvat.apps.authentication.decorators import login_required
-from cvat.apps.engine.models import Task as TaskModel
-from cvat.apps.engine import annotation
-
 import django_rq
 import fnmatch
 import json
 import os
 import rq
 
+from django.http import HttpResponse, JsonResponse, HttpResponseBadRequest
+from django.db.models import Q
+from rules.contrib.views import permission_required, objectgetter
+
+from cvat.apps.authentication.decorators import login_required
+from cvat.apps.engine.models import Task as TaskModel
+from cvat.apps.engine import annotation
+from cvat.apps.authentication.auth import has_admin_role
 from cvat.apps.engine.log import slogger
 
 from .model_loader import ModelLoader, load_label_map
 from .image_loader import ImageLoader
-import os.path as osp
+from . import model_manager
+from .models import AnnotationModel
 
 def get_image_data(path_to_data):
     def get_image_key(item):
-        return int(osp.splitext(osp.basename(item))[0])
+        return int(os.path.splitext(os.path.basename(item))[0])
 
     image_list = []
     for root, _, filenames in os.walk(path_to_data):
         for filename in fnmatch.filter(filenames, "*.jpg"):
-                image_list.append(osp.join(root, filename))
+                image_list.append(os.path.join(root, filename))
 
     image_list.sort(key=get_image_key)
     return ImageLoader(image_list)
@@ -202,7 +205,6 @@ def run_inference_engine_annotation(path_to_data, model_file, weights_file,
         frame_counter += 1
         if not update_progress(job, frame_counter * 100 / data_len):
             return None
-
     processed_detections = process_detections(detections, convertation_file)
 
     add_boxes(processed_detections.get_boxes(), result["create"]["boxes"])
@@ -222,7 +224,7 @@ def update_progress(job, progress):
     job.save_meta()
     return True
 
-def create_thread(tid, model_file, weights_file, labels_mapping, attributes, convertation_file):
+def create_thread(tid, model_file, weights_file, labels_mapping, attributes, convertation_file, reset):
     try:
         job = rq.get_current_job()
         job.meta["progress"] = 0
@@ -246,137 +248,18 @@ def create_thread(tid, model_file, weights_file, labels_mapping, attributes, con
             slogger.glob.info("auto annotation for task {} canceled by user".format(tid))
             return
 
-        annotation.clear_task(tid)
+        if reset:
+            annotation.clear_task(tid)
         annotation.save_task(tid, result)
         slogger.glob.info("auto annotation for task {} done".format(tid))
-    except Exception:
+    except Exception as e:
         try:
             slogger.task[tid].exception("exception was occurred during auto annotation of the task", exc_info=True)
         except Exception as ex:
             slogger.glob.exception("exception was occurred during auto annotation of the task {}: {}".format(tid, str(ex)), exc_info=True)
+            raise ex
 
-@login_required
-def get_meta_info(request):
-    try:
-        queue = django_rq.get_queue("low")
-        tids = json.loads(request.body.decode("utf-8"))
-        result = {}
-        for tid in tids:
-            job = queue.fetch_job("auto_annotation.create/{}".format(tid))
-            if job is not None:
-                result[tid] = {
-                    "active": job.is_queued or job.is_started,
-                    "success": not job.is_failed
-                }
-
-        return JsonResponse(result)
-    except Exception as ex:
-        slogger.glob.exception("exception was occurred during auto annotation meta request", exc_info=True)
-        return HttpResponseBadRequest(str(ex))
-
-@login_required
-@permission_required(perm=["engine.task.change"],
-    fn=objectgetter(TaskModel, "tid"), raise_exception=True)
-def create(request, tid):
-    slogger.glob.info("auto annotation create request for task {}".format(tid))
-
-    def write_file(path, file_obj):
-        with open(path, "wb") as upload_file:
-            for chunk in file_obj.chunks():
-                upload_file.write(chunk)
-
-    try:
-        db_task = TaskModel.objects.get(pk=tid)
-        upload_dir = db_task.get_upload_dirname()
-        queue = django_rq.get_queue("low")
-        job = queue.fetch_job("auto_annotation.create/{}".format(tid))
-        if job is not None and (job.is_started or job.is_queued):
-            raise Exception("The process is already running")
-
-        model_file = request.FILES["model"]
-        model_file_path = os.path.join(upload_dir, model_file.name)
-        write_file(model_file_path, model_file)
-
-        weights_file = request.FILES["weights"]
-        weights_file_path = os.path.join(upload_dir, weights_file.name)
-        write_file(weights_file_path, weights_file)
-
-        config_file = request.FILES["config"]
-        config_file_path = os.path.join(upload_dir, config_file.name)
-        write_file(config_file_path, config_file)
-
-        convertation_file = request.FILES["conv_script"]
-        convertation_file_path = os.path.join(upload_dir, convertation_file.name)
-        write_file(convertation_file_path, convertation_file)
-
-        db_labels = db_task.label_set.prefetch_related("attributespec_set").all()
-        db_attributes = {db_label.id:
-            {db_attr.get_name(): db_attr.id for db_attr in db_label.attributespec_set.all()} for db_label in db_labels}
-        db_labels = {db_label.id:db_label.name for db_label in db_labels}
-
-        class_names = load_label_map(config_file_path)
-
-        labels_mapping = {}
-        for db_key, db_label in db_labels.items():
-            for key, label in class_names.items():
-                if label == db_label:
-                    labels_mapping[int(key)] = db_key
-
-        if not labels_mapping:
-            raise Exception("No labels found for annotation")
-
-        queue.enqueue_call(func=create_thread,
-            args=(
-                tid,
-                model_file_path,
-                weights_file_path,
-                labels_mapping,
-                db_attributes,
-                convertation_file_path),
-            job_id="auto_annotation.create/{}".format(tid),
-            timeout=604800)     # 7 days
-
-        slogger.task[tid].info("auto annotation job enqueued")
-
-    except Exception as ex:
-        try:
-            slogger.task[tid].exception("exception was occurred during annotation request", exc_info=True)
-        except Exception as logger_ex:
-            slogger.glob.exception("exception was occurred during create auto annotation request for task {}: {}".format(tid, str(logger_ex)), exc_info=True)
-        return HttpResponseBadRequest(str(ex))
-
-    return HttpResponse()
-
-@login_required
-@permission_required(perm=["engine.task.access"],
-    fn=objectgetter(TaskModel, "tid"), raise_exception=True)
-def check(request, tid):
-    try:
-        queue = django_rq.get_queue("low")
-        job = queue.fetch_job("auto_annotation.create/{}".format(tid))
-        if job is not None and "cancel" in job.meta:
-            return JsonResponse({"status": "finished"})
-        data = {}
-        if job is None:
-            data["status"] = "unknown"
-        elif job.is_queued:
-            data["status"] = "queued"
-        elif job.is_started:
-            data["status"] = "started"
-            data["progress"] = job.meta["progress"]
-        elif job.is_finished:
-            data["status"] = "finished"
-            job.delete()
-        else:
-            data["status"] = "failed"
-            data["stderr"] = job.exc_info
-            job.delete()
-
-    except Exception:
-        data["status"] = "unknown"
-
-    return JsonResponse(data)
-
+        raise e
 
 @login_required
 @permission_required(perm=["engine.task.change"],
@@ -384,7 +267,7 @@ def check(request, tid):
 def cancel(request, tid):
     try:
         queue = django_rq.get_queue("low")
-        job = queue.fetch_job("auto_annotation.create/{}".format(tid))
+        job = queue.fetch_job("auto_annotation.run.{}".format(tid))
         if job is None or job.is_finished or job.is_failed:
             raise Exception("Task is not being annotated currently")
         elif "cancel" not in job.meta:
@@ -399,3 +282,220 @@ def cancel(request, tid):
         return HttpResponseBadRequest(str(ex))
 
     return HttpResponse()
+
+
+@login_required
+@permission_required(perm=["auto_annotation.model.create"], raise_exception=True)
+def create_model(request):
+    if request.method != 'POST':
+        return HttpResponseBadRequest("Only POST requests are accepted")
+
+    try:
+        params = request.POST
+        storage = params["storage"]
+        name = params["name"]
+        is_shared = params["shared"].lower() == "true"
+        if is_shared and not has_admin_role(request.user):
+            raise Exception("Only admin can create shared models")
+
+        files = request.FILES if storage == "local" else params
+        model = files["xml"]
+        weights = files["bin"]
+        labelmap = files["json"]
+        interpretation_script = files["py"]
+        owner = request.user
+
+        dl_model_id = model_manager.create_empty(owner=owner)
+        rq_id = model_manager.update_model(
+            dl_model_id=dl_model_id,
+            name=name,
+            model_file=model,
+            weights_file=weights,
+            labelmap_file=labelmap,
+            interpretation_file=interpretation_script,
+            storage=storage,
+            is_shared=is_shared,
+        )
+
+        return JsonResponse({"id": rq_id})
+    except Exception as e:
+        return HttpResponseBadRequest(str(e))
+
+@login_required
+@permission_required(perm=["auto_annotation.model.update"],
+    fn=objectgetter(AnnotationModel, "mid"), raise_exception=True)
+def update_model(request, mid):
+    if request.method != 'POST':
+        return HttpResponseBadRequest("Only POST requests are accepted")
+
+    try:
+        params = request.POST
+        storage = params["storage"]
+        name = params.get("name")
+        is_shared = params.get("shared")
+        is_shared = is_shared.lower() == "true" if is_shared else None
+        if is_shared and not has_admin_role(request.user):
+            raise Exception("Only admin can create shared models")
+        files = request.FILES
+        model = files.get("xml")
+        weights = files.get("bin")
+        labelmap = files.get("json")
+        interpretation_script = files.get("py")
+
+        rq_id = model_manager.update_model(
+            dl_model_id=mid,
+            name=name,
+            model_file=model,
+            weights_file=weights,
+            labelmap_file=labelmap,
+            interpretation_file=interpretation_script,
+            storage=storage,
+            is_shared=is_shared,
+        )
+
+        return JsonResponse({"id": rq_id})
+    except Exception as e:
+        return HttpResponseBadRequest(str(e))
+
+@login_required
+@permission_required(perm=["auto_annotation.model.delete"],
+    fn=objectgetter(AnnotationModel, "mid"), raise_exception=True)
+def delete_model(request, mid):
+    if request.method != 'DELETE':
+        return HttpResponseBadRequest("Only DELETE requests are accepted")
+    model_manager.delete(mid)
+    return HttpResponse()
+
+@login_required
+def get_meta_info(request):
+    try:
+        tids = json.loads(request.body.decode('utf-8'))
+        response = {
+            "admin": has_admin_role(request.user),
+            "models": [],
+            "run": {},
+        }
+        dl_model_list = list(AnnotationModel.objects.filter(Q(owner=request.user) | Q(primary=True) | Q(shared=True)).order_by('-created_date'))
+        for dl_model in dl_model_list:
+            labels = []
+            if dl_model.labelmap_file and os.path.exists(dl_model.labelmap_file.name):
+                with dl_model.labelmap_file.open('r') as f:
+                    labels = list(json.load(f)["label_map"].values())
+
+            response["models"].append({
+                "id": dl_model.id,
+                "name": dl_model.name,
+                "primary": dl_model.primary,
+                "uploadDate": dl_model.created_date,
+                "updateDate": dl_model.updated_date,
+                "labels": labels,
+            })
+
+        queue = django_rq.get_queue("low")
+        for tid in tids:
+            rq_id = "auto_annotation.run.{}".format(tid)
+            job = queue.fetch_job(rq_id)
+            if job is not None:
+                response["run"][tid] = {
+                    "status": job.status,
+                    "rq_id": rq_id,
+                }
+
+        return JsonResponse(response)
+    except Exception as e:
+        return HttpResponseBadRequest(str(e))
+
+@login_required
+@permission_required(perm=["engine.task.change"],
+    fn=objectgetter(TaskModel, "tid"), raise_exception=True)
+@permission_required(perm=["auto_annotation.model.access"],
+    fn=objectgetter(AnnotationModel, "mid"), raise_exception=True)
+def start_annotation(request, mid, tid):
+    slogger.glob.info("auto annotation create request for task {} via DL model {}".format(tid, mid))
+    try:
+        db_task = TaskModel.objects.get(pk=tid)
+        queue = django_rq.get_queue("low")
+        job = queue.fetch_job("auto_annotation.run.{}".format(tid))
+        if job is not None and (job.is_started or job.is_queued):
+            raise Exception("The process is already running")
+
+        data = json.loads(request.body.decode('utf-8'))
+
+        should_reset = data["reset"]
+        user_defined_labels_mapping = data["labels"]
+
+        dl_model = AnnotationModel.objects.get(pk=mid)
+
+        model_file_path = dl_model.model_file.name
+        weights_file_path = dl_model.weights_file.name
+        labelmap_file = dl_model.labelmap_file.name
+        convertation_file_path = dl_model.interpretation_file.name
+
+        db_labels = db_task.label_set.prefetch_related("attributespec_set").all()
+        db_attributes = {db_label.id:
+            {db_attr.get_name(): db_attr.id for db_attr in db_label.attributespec_set.all()} for db_label in db_labels}
+        db_labels = {db_label.name:db_label.id for db_label in db_labels}
+
+        model_labels = {value: key for key, value in load_label_map(labelmap_file).items()}
+
+        labels_mapping = {}
+        for user_model_label, user_db_label in user_defined_labels_mapping.items():
+            if user_model_label in model_labels and user_db_label in db_labels:
+                labels_mapping[int(model_labels[user_model_label])] = db_labels[user_db_label]
+
+        if not labels_mapping:
+            raise Exception("No labels found for annotation")
+
+        rq_id="auto_annotation.run.{}".format(tid)
+        queue.enqueue_call(func=create_thread,
+            args=(
+                tid,
+                model_file_path,
+                weights_file_path,
+                labels_mapping,
+                db_attributes,
+                convertation_file_path,
+                should_reset,
+            ),
+            job_id = rq_id,
+            timeout=604800)     # 7 days
+
+        slogger.task[tid].info("auto annotation job enqueued")
+
+    except Exception as ex:
+        try:
+            slogger.task[tid].exception("exception was occurred during annotation request", exc_info=True)
+        except Exception as logger_ex:
+            slogger.glob.exception("exception was occurred during create auto annotation request for task {}: {}".format(tid, str(logger_ex)), exc_info=True)
+        return HttpResponseBadRequest(str(ex))
+
+    return JsonResponse({"id": rq_id})
+
+@login_required
+def check(request, rq_id):
+    try:
+        target_queue = "low" if "auto_annotation.run" in rq_id else "default"
+        queue = django_rq.get_queue(target_queue)
+        job = queue.fetch_job(rq_id)
+        if job is not None and "cancel" in job.meta:
+            return JsonResponse({"status": "finished"})
+        data = {}
+        if job is None:
+            data["status"] = "unknown"
+        elif job.is_queued:
+            data["status"] = "queued"
+        elif job.is_started:
+            data["status"] = "started"
+            data["progress"] = job.meta["progress"] if "progress" in job.meta else ""
+        elif job.is_finished:
+            data["status"] = "finished"
+            job.delete()
+        else:
+            data["status"] = "failed"
+            data["error"] = job.exc_info
+            job.delete()
+
+    except Exception:
+        data["status"] = "unknown"
+
+    return JsonResponse(data)