+++ cvat/apps/dashboard/static/dashboard/js/dashboard.js
@@ -4,6 +4,18 @@
  * SPDX-License-Identifier: MIT
  */
 
+/* global
+    AnnotationParser:false
+    Config:false
+    userConfirm:false
+    ConstIdGenerator:false
+    createExportContainer:false
+    dumpAnnotationRequest:false
+    LabelsInfo:false
+    showMessage:false
+    showOverlay:false
+*/
+
 "use strict";
 /* Server requests */
 function createTaskRequest(oData, onSuccessRequest, onSuccessCreate, onError, onComplete, onUpdateStatus) {
@@ -85,7 +97,7 @@ function updateTaskRequest(labels) {
 
 
 function removeTaskRequest() {
-    confirm("The action can not be undone. Are you sure?", confirmCallback);
+    userConfirm("The action can not be undone. Are you sure?", confirmCallback);
 
     function confirmCallback() {
         $.ajax ({
@@ -254,7 +266,7 @@ window.cvat.dashboard.uiCallbacks.push(function(elements) {
         uploadButton.on("click", function() {
             window.cvat.dashboard.taskID = taskID;
             window.cvat.dashboard.taskName = taskName;
-            confirm("The current annotation will be lost. Are you sure?", uploadAnnotationRequest);
+            userConfirm("The current annotation will be lost. Are you sure?", uploadAnnotationRequest);
         });
 
         updateButton.on("click", function() {
@@ -388,9 +400,11 @@ function setupTaskCreator() {
 
     cancelBrowseServer.on("click", () => shareFileSelector.addClass("hidden"));
     submitBrowseServer.on("click", function() {
-        files = shareBrowseTree.jstree(true).get_selected();
-        cancelBrowseServer.click();
-        updateSelectedFiles();
+        if (!createModal.hasClass("hidden")) {
+            files = shareBrowseTree.jstree(true).get_selected();
+            cancelBrowseServer.click();
+            updateSelectedFiles();
+        }
     });
 
     flipImagesBox.on("click", (e) => {