+++ cvat/apps/documentation/static/documentation/js/shortcuts.js
@@ -4,6 +4,10 @@
  * SPDX-License-Identifier: MIT
  */
 
+/* global
+    Mousetrap:false
+*/
+
 Mousetrap.bind(window.cvat.config.shortkeys["open_help"].value, function() {
     window.open("/documentation/user_guide.html");