+++ cvat/apps/engine/static/engine/js/player.js
@@ -5,6 +5,15 @@
  */
 
 /* exported PlayerModel PlayerController PlayerView */
+
+/* global
+    blurAllElements:false
+    copyToClipboard:false
+    Listener:false
+    Logger:false
+    Mousetrap:false
+*/
+
 "use strict";
 
 class FrameProvider extends Listener {
@@ -493,6 +502,7 @@ class PlayerController {
     }
 
     fit() {
+        Logger.addEvent(Logger.EventType.fitImage)
         this._model.fit();
     }